GAP-100: ASTERIX Category 205 direction-finder bearings
Pins EUROCONTROL-SPEC-0149-31 edition 1.0 (docs/design/external-standards.md
§9, fetched and read in full) and adds gungnir-interop/src/asterix/cat205.rs:
a lossless decoder for every standard-UAP item, with the three items edition
1.0 itself calls "implementation dependent" (I205/100, /120, /170) carried
raw and named, the same treatment Category 048 already gives items it does
not interpret.

The one real design question -- this category states no angular error for a
bearing anywhere, so a variance cannot honestly be read off the wire -- is
answered the same way gungnir_ingest::adapters::sapient already answers it
for SAPIENT: DfSite::azimuth_sigma_rad is the deployment's own stated
accuracy from that direction finder's ICD, supplied by the caller and never
invented. Confirmed gungnir_model::Measurement::Bearing (DN-27, GAP-096)
unchanged on main before reusing it, per the task's own instruction to check
rather than assume.

AsterixFeedAdapter gains a third category arm and an opt-in with_df_sites
builder, so no existing call site changes. Message types 1/3 (the RDF
processing system's own already-resolved position) decode losslessly but are
deliberately not mapped: converting them needs gungnir-geo, which
gungnir-interop may not depend on. Category 129 (UAS Identification
Reports), surveyed alongside 205, is deliberately not built here -- a
different report shape that would not share this gap's one hard question.

No real Category 205 capture exists to vendor (checked EUROCONTROL,
CroatiaControlLtd/asterix, and asterix-specs); testdata/asterix/cat205.raw is
hand-built from the specification's own byte tables and documented as
exactly that in testdata/asterix/SOURCE.md, never presented as a real-world
recording.

Renumbered from GAP-099 to GAP-100 during merge: GAP-099 was already taken
by the MISB ST 0601 adapter (PR #45), filed independently around the same
time -- the same numbering-collision pattern this register's own history
already documents happening repeatedly.

Human-owned crate touched (gungnir-ingest, the low-trust gateway); written
and gated, not signed.

+++ gungnir-interop/src/asterix/cat048.rs
@@ -363,7 +363,7 @@ fn parse_record(cur: &mut Cursor<'_>) -> Result<Record, InteropError> {
                 r.flight_level = Some(FlightLevel {
                     validated: v & 0x8000 == 0,
                     garbled: v & 0x4000 != 0,
-                    level: f64::from(sign_extend(v & 0x3FFF, 14)) / 4.0,
+                    level: f64::from(sign_extend(u32::from(v) & 0x3FFF, 14)) / 4.0,
                 });
             }
             7 => r.plot_characteristics = Some(parse_plot_characteristics(cur)?),
@@ -436,7 +436,7 @@ fn parse_record(cur: &mut Cursor<'_>) -> Result<Record, InteropError> {
             18 => raw(&mut r, "I048/100", cur.take(4, "I048/100")?),
             19 => {
                 let v = cur.u16("I048/110")?;
-                r.height_3d_ft = Some(f64::from(sign_extend(v & 0x3FFF, 14)) * 25.0);
+                r.height_3d_ft = Some(f64::from(sign_extend(u32::from(v) & 0x3FFF, 14)) * 25.0);
             }
             20 => r.doppler = Some(parse_doppler(cur)?),
             21 => raw(&mut r, "I048/230", cur.take(2, "I048/230")?),
@@ -515,7 +515,7 @@ fn parse_doppler(cur: &mut Cursor<'_>) -> Result<RadialDoppler, InteropError> {
     let mut d = RadialDoppler::default();
     if p & 0x80 != 0 {
         let v = cur.u16("I048/120 CAL")?;
-        d.calculated = Some((v & 0x8000 != 0, sign_extend(v & 0x03FF, 10)));
+        d.calculated = Some((v & 0x8000 != 0, sign_extend(u32::from(v) & 0x03FF, 10)));
     }
     if p & 0x40 != 0 {
         let rep = cur.u8("I048/120 REP")?;
